Move listLeafFiles() to InMemoryFileIndex
